+++ changelog.md
@@ -5,3 +5,7 @@
 - Init
 - Added product extraction
 - Added sitemap products url extraction
+- Renamed proc that gets product urls
+- Added tests
+- Fixed crash at non Shopify-made pages
+- Fixed product urls extraction, removed non product urls